rebase on main & remove duplicated component code from views
