+++ server/referralDetails/availabilityView.ts
@@ -1,25 +1,8 @@
-import { ButtonArgs, InsetTextArgs, SummaryListArgs } from '../utils/govukFrontendTypes'
-import ViewUtils from '../utils/viewUtils'
+import { ButtonArgs, InsetTextArgs } from '../utils/govukFrontendTypes'
 import AvailabilityPresenter from './availabilityPresenter'
 
 export default class AvailabilityView {
   constructor(private readonly presenter: AvailabilityPresenter) {}
-
-  get summary(): SummaryListArgs {
-    return {
-      ...ViewUtils.summaryListArgsWithSummaryCard(this.presenter.personalDetailsSummaryList(), 'Personal details'),
-    }
-  }
-
-  get referralSummary(): SummaryListArgs {
-    return {
-      ...ViewUtils.summaryListArgs(
-        this.presenter.referralSummaryList(),
-        { showBorders: false },
-        'govuk-!-margin-bottom-0',
-      ),
-    }
-  }
 
   get importFromDeliusText(): InsetTextArgs {
     return {
@@ -40,8 +23,6 @@ export default class AvailabilityView {
       'referralDetails/referralDetails',
       {
         presenter: this.presenter,
-        summary: this.summary,
-        referralSummary: this.referralSummary,
         importFromDeliusText: this.importFromDeliusText,
         availabilityButtonArgs: this.availabilityButtonArgs,
       },
