+++ server/referralDetails/locationView.ts
@@ -1,33 +1,13 @@
-import { SummaryListArgs } from '../utils/govukFrontendTypes'
-import ViewUtils from '../utils/viewUtils'
-import AvailabilityPresenter from './availabilityPresenter'
+import LocationPresenter from './locationPresenter'
 
 export default class LocationView {
-  constructor(private readonly presenter: AvailabilityPresenter) {}
-
-  get summary(): SummaryListArgs {
-    return {
-      ...ViewUtils.summaryListArgsWithSummaryCard(this.presenter.personalDetailsSummaryList(), 'Personal details'),
-    }
-  }
-
-  get referralSummary(): SummaryListArgs {
-    return {
-      ...ViewUtils.summaryListArgs(
-        this.presenter.referralSummaryList(),
-        { showBorders: false },
-        'govuk-!-margin-bottom-0',
-      ),
-    }
-  }
+  constructor(private readonly presenter: LocationPresenter) {}
 
   get renderArgs(): [string, Record<string, unknown>] {
     return [
       'referralDetails/referralDetails',
       {
         presenter: this.presenter,
-        summary: this.summary,
-        referralSummary: this.referralSummary,
       },
     ]
   }
