+++ server/referralDetails/referralDetailsPresenter.ts
@@ -1,6 +1,8 @@
 import PersonalDetails from '../models/PersonalDetails'
 import { SummaryListItem } from '../utils/summaryList'
 import DateUtils from '../utils/dateUtils'
+import { SummaryListArgs } from '../utils/govukFrontendTypes'
+import ViewUtils from '../utils/viewUtils'
 
 export enum ReferralDetailsPageSection {
   PersonalDetailsTab = 'personalDetails',
@@ -18,6 +20,12 @@ export default class ReferralDetailsPresenter {
     readonly subNavValue: string,
     readonly id: string,
   ) {}
+
+  get referralSummary(): SummaryListArgs {
+    return {
+      ...ViewUtils.summaryListArgs(this.referralSummaryList(), { showBorders: false }, 'govuk-!-margin-bottom-0'),
+    }
+  }
 
   getSubHeaderArgs(): {
     heading: { text: string; classes: string }
@@ -174,61 +182,5 @@ export default class ReferralDetailsPresenter {
         valueLink: '<a href="mailto:tom.saunders@justice.gov.uk">tom.saunders@justice.gov.uk</a>',
       },
     ]
-  }
-
-  offenceHistorySummaryLists(): { title: string; summary: SummaryListItem[] }[] {
-    const summaries = []
-    summaries.push({
-      title: 'Index Offence',
-      summary: [
-        {
-          key: 'Offence',
-          lines: ['Publishing or causing to be published a tobacco advertisement - 09144'],
-        },
-        {
-          key: 'Category',
-          lines: ['Public Health Offences'],
-        },
-        {
-          key: 'Offence date',
-          lines: ['11 June 2020'],
-        },
-      ],
-    })
-    summaries.push({
-      title: 'Additional offence (08000)',
-      summary: [
-        {
-          key: 'Offence',
-          lines: ['Absconding from lawful custody - 08000'],
-        },
-        {
-          key: 'Category',
-          lines: ['Absconding from lawful custody'],
-        },
-        {
-          key: 'Offence date',
-          lines: ['18 January 2013'],
-        },
-      ],
-    })
-    summaries.push({
-      title: 'Additional offence (08000)',
-      summary: [
-        {
-          key: 'Offence',
-          lines: ['\tClass unspecified - permitting premises to be used - 09329'],
-        },
-        {
-          key: 'Category',
-          lines: ['Permitting premises to be used for unlawful (drug-related) purposes'],
-        },
-        {
-          key: 'Offence date',
-          lines: ['23 September 2000'],
-        },
-      ],
-    })
-    return summaries
   }
 }
