+++ server/views/referralDetails/availability.njk
@@ -6,6 +6,5 @@
    {{ govukInsetText(importFromDeliusText) }}
    <p>No availability details added for Alex River</p>
 
-   {{ govukButton(availabilityButtonArgs) }}
+<!--   {{ govukButton(availabilityButtonArgs) }}-->
 </div>
-
