+++ server/views/referralDetails/referralDetails.njk
@@ -26,7 +26,7 @@
         {{ mojSubNavigation(presenter.getHorizontalSubNavArgs()) }}
         <div class="govuk-grid-column-three-quarters">
           <h2 class="govuk-heading-m">Referral summary</h2>
-          {{ govukSummaryList(referralSummary) }}
+          {{ govukSummaryList(presenter.referralSummary) }}
         </div>
       </div>
     </div>
